requirements.txt
requests~=2.32.5
python-dotenv~=1.2.1
pyahocorasick~=2.2.0
django-emoji~=2.2.2
python-decouple~=3.8